+++ packages/math/types/integers.d.ts
@@ -8,7 +8,16 @@ interface WithByteConverters {
   readonly toBytesLittleEndian: () => Uint8Array;
 }
 export declare class Uint32 implements Integer, WithByteConverters {
+  /** @deprecated use Uint32.fromBytes */
   static fromBigEndianBytes(bytes: ArrayLike<number>): Uint32;
+  /**
+   * Creates a Uint32 from a fixed length byte array.
+   *
+   * @param bytes a list of exactly 4 bytes
+   * @param endianess defaults to big endian
+   */
+  static fromBytes(bytes: ArrayLike<number>, endianess?: "be" | "le"): Uint32;
+  static fromString(str: string): Uint32;
   protected readonly data: number;
   constructor(input: number);
   toBytesBigEndian(): Uint8Array;
@@ -31,7 +40,15 @@ export declare class Uint53 implements Integer {
   toString(): string;
 }
 export declare class Uint64 implements Integer, WithByteConverters {
+  /** @deprecated use Uint64.fromBytes */
   static fromBytesBigEndian(bytes: ArrayLike<number>): Uint64;
+  /**
+   * Creates a Uint64 from a fixed length byte array.
+   *
+   * @param bytes a list of exactly 8 bytes
+   * @param endianess defaults to big endian
+   */
+  static fromBytes(bytes: ArrayLike<number>, endianess?: "be" | "le"): Uint64;
   static fromString(str: string): Uint64;
   static fromNumber(input: number): Uint64;
   private readonly data;